refactor(wizards/fcda): move finder-list definition
